refactor(markdown-to-ast): dir up
