--- packages/markdown-to-ast/lib/markdown/union-syntax.js
@@ -1,15 +1,0 @@
-// LICENSE : MIT
-"use strict";
-
-var _exports;
-
-function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
-
-var _require = require("@textlint/ast-node-types"),
-    ASTNodeTypes = _require.ASTNodeTypes;
-
-var T = ASTNodeTypes;
-// public key interface
-var _exports2 = (_exports = {}, _defineProperty(_exports, T.Document, T.Document), _defineProperty(_exports, T.Paragraph, T.Paragraph), _defineProperty(_exports, T.BlockQuote, T.BlockQuote), _defineProperty(_exports, T.ListItem, T.ListItem), _defineProperty(_exports, T.List, T.List), _defineProperty(_exports, T.Header, T.Header), _defineProperty(_exports, T.CodeBlock, T.CodeBlock), _defineProperty(_exports, T.HtmlBlock, T.HtmlBlock), _defineProperty(_exports, T.ReferenceDef, T.ReferenceDef), _defineProperty(_exports, T.HorizontalRule, T.HorizontalRule), _defineProperty(_exports, T.Str, T.Str), _defineProperty(_exports, T.Break, T.Break), _defineProperty(_exports, T.Emphasis, T.Emphasis), _defineProperty(_exports, T.Strong, T.Strong), _defineProperty(_exports, T.Html, T.Html), _defineProperty(_exports, T.Link, T.Link), _defineProperty(_exports, T.Image, T.Image), _defineProperty(_exports, T.Code, T.Code), _exports);
-module.exports = _exports2;
-//# sourceMappingURL=union-syntax.js.map
